Run static site build preview on all branches and PRs
A while back, our TravisCI was set up in a way where it deployed the
static site. However, Travis was configured so that no jobs ran for
branches or pull requests, causing those builds to wait in GitHub on a
build that would never come.

This change sets up a simple build of the static site contents to
provide some assurances that changes are valid, and ensures its run on
all branches and pull requests.

Fixes https://github.com/pelias/pelias/issues/841